bazar/requirements.txt
appdirs==1.4.3
Django==1.11.1
django-bootstrap3==8.2.3
django-chartjs==1.2
django-inline-svg==0.1.1
django-pyowm==0.2.0
djangorestframework==3.6.3
packaging==16.8
pyowm==2.6.1
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
django-rest-swagger==2.2.0
